AutoDL Thumb Adjustment
Auto Download will not overwrite a thumb if it currently exists on the
File/Post

+++ ts_filecabinet.php
@@ -845,8 +845,11 @@ function tsfc_save_custom_metabox( $post_id ) {
 		$type = tsfc_video_type($post_id);
 		if ($type == 'vimeo_video' || $type == 'youtube_video'){
 			$post_thumbnail_id = get_post_thumbnail_id( $post_id );
-			wp_delete_post( $post_thumbnail_id, true );
-			tsfc_check_add_thumb($post_id);
+			//if post has a thumbnail is empty we autoDL else we do nothing		
+			if (empty($post_thumbnail_id)){
+				wp_delete_post( $post_thumbnail_id, true );
+				tsfc_check_add_thumb($post_id);
+			}
 		}
 	}
